move example_dir to pytest_helpers
