+++ pytest_ipynb2/pytester_helpers.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from textwrap import indent
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Protocol
 
 import pytest
 
@@ -190,3 +190,27 @@ class ExampleDir:
 
     files: list[Path]
     conftest: str = ""
+
+
+class ExampleDirRequest(Protocol):
+    """Typehint to param passed to example_dir."""
+
+    param: ExampleDir
+
+
+@pytest.fixture
+def example_dir(request: ExampleDirRequest, pytester: pytest.Pytester) -> CollectedDir:
+    """Parameterised fixture. Requires a list of `Path`s to copy into a pytester instance."""
+    if request.param.conftest:
+        pytester.makeconftest(request.param.conftest)
+
+    for filetocopy in request.param.files:
+        pytester.copy_example(str(filetocopy))
+
+    dir_node = pytester.getpathnode(pytester.path)
+
+    return CollectedDir(
+        pytester_instance=pytester,
+        dir_node=dir_node,
+        items=pytester.genitems([dir_node]),
+    )
