+++ tests/test_collection.py
@@ -1,11 +1,10 @@
 from pathlib import Path
 
 import pytest
-from conftest import CollectedDir
 
 import pytest_ipynb2
 import pytest_ipynb2.plugin
-from pytest_ipynb2.pytester_helpers import CollectionTree, ExampleDir
+from pytest_ipynb2.pytester_helpers import CollectedDir, CollectionTree, ExampleDir
 
 
 @pytest.fixture
